requirements.txt
openai>=1.59.0
pydantic>=2.6.0
python-dotenv>=1.0.0
